python-service/pyproject.toml
[tool.poetry]
name = "macos-voice-assistant"
version = "1.0.0"
description = "Intelligent privacy-first voice assistant for macOS Tahoe 26.1"
authors = ["Voice Assistant Contributors"]
license = "Apache-2.0"
readme = "README.md"
homepage = "https://github.com/yourusername/macos-voice-assistant"
repository = "https://github.com/yourusername/macos-voice-assistant"
keywords = ["voice-assistant", "macos", "ai", "siri-alternative", "privacy"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "License :: OSI Approved :: Apache Software License",
    "Operating System :: MacOS :: MacOS X",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Multimedia :: Sound/Audio :: Speech",
]

[tool.poetry.dependencies]
python = "^3.9"

# macOS Integration
pyobjc-framework-Cocoa = "^11.0"
pyobjc-framework-ApplicationServices = "^11.0"
pyobjc-framework-AVFoundation = "^11.0"
pyobjc-framework-CoreAudio = "^11.0"

# Wake Word Detection
pvporcupine = "^3.0.0"

# Audio Processing
pyaudio = "^0.2.14"
soundfile = "^0.12.1"
numpy = "^1.24.0"
scipy = "^1.11.0"

# Speech Recognition (Whisper dependencies)
torch = "^2.0.0"
torchaudio = "^2.0.0"

# Voice Activity Detection
silero-vad = "^5.1"

# LLM Clients
mlx-lm = "^0.7.0"  # For local gpt-oss on Apple Silicon
openai = "^1.0.0"
anthropic = "^0.25.0"
httpx = "^0.27.0"

# MCP Server
fastmcp = "^2.0.0"
mcp = "^1.0.0"

# Configuration & Data
pyyaml = "^6.0"
pydantic = "^2.0.0"
python-dotenv = "^1.0.0"

# Async & Concurrency
asyncio = "^3.4.3"
aiofiles = "^23.0.0"

# Retry & Error Handling
tenacity = "^8.2.0"

# Logging & Monitoring
loguru = "^0.7.0"
psutil = "^5.9.0"

# Security (Keychain access)
keyring = "^24.0.0"

# Web Search (for MCP tool)
duckduckgo-search = "^4.0.0"

# CLI utilities
click = "^8.1.0"
rich = "^13.0.0"

[tool.poetry.group.dev.dependencies]
# Testing
pytest = "^7.0.0"
pytest-asyncio = "^0.21.0"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
pytest-timeout = "^2.2.0"

# Code Quality
black = "^23.0.0"
isort = "^5.13.0"
ruff = "^0.1.0"
mypy = "^1.0.0"

# Type Stubs
types-PyYAML = "^6.0.0"
types-requests = "^2.31.0"

# Profiling & Debugging
ipython = "^8.0.0"
ipdb = "^0.13.0"
memory-profiler = "^0.61.0"
py-spy = "^0.3.0"

# Documentation
sphinx = "^7.0.0"
sphinx-rtd-theme = "^2.0.0"

[tool.poetry.scripts]
voice-assistant = "voice_assistant.main:main"
voice-assistant-setup = "voice_assistant.setup:setup"
voice-assistant-test = "voice_assistant.test_utils:test_audio"

[tool.poetry.extras]
elevenlabs = ["elevenlabs"]
all = ["elevenlabs"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

# Black Configuration
[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

# isort Configuration
[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true

# Pytest Configuration
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py", "*_test.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-ra",
    "--strict-markers",
    "--strict-config",
    "--cov=voice_assistant",
    "--cov-report=term-missing:skip-covered",
    "--cov-report=html",
    "--cov-report=xml",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: marks tests as integration tests",
    "unit: marks tests as unit tests",
    "macos: marks tests that require macOS",
]
asyncio_mode = "auto"

# MyPy Configuration
[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "pvporcupine.*",
    "pyaudio.*",
    "silero_vad.*",
]
ignore_missing_imports = true

# Ruff Configuration
[tool.ruff]
line-length = 100
target-version = "py39"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

# Coverage Configuration
[tool.coverage.run]
branch = true
source = ["voice_assistant"]
omit = [
    "*/tests/*",
    "*/test_*.py",
    "*/__pycache__/*",
    "*/venv/*",
    "*/virtualenv/*",
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]
